beginning of reorganizing of test Dummy Data
